Export `AddExtension` from `middleware` (#811)
* Export `AddExtension` from `middleware`

* Move `AddExtension` into `extension` module
