+++ axum/src/extension.rs
@@ -1,8 +1,15 @@
-use crate::extract::{rejection::*, FromRequest, RequestParts};
-use crate::response::IntoResponseParts;
+use crate::{
+    extract::{rejection::*, FromRequest, RequestParts},
+    response::IntoResponseParts,
+};
 use async_trait::async_trait;
 use axum_core::response::{IntoResponse, Response, ResponseParts};
-use std::ops::Deref;
+use http::Request;
+use std::{
+    ops::Deref,
+    task::{Context, Poll},
+};
+use tower_service::Service;
 
 /// Extractor and response for extensions.
 ///
@@ -120,12 +127,44 @@ impl<S, T> tower_layer::Layer<S> for Extension<T>
 where
     T: Clone + Send + Sync + 'static,
 {
-    type Service = crate::AddExtension<S, T>;
+    type Service = AddExtension<S, T>;
 
     fn layer(&self, inner: S) -> Self::Service {
-        crate::AddExtension {
+        AddExtension {
             inner,
             value: self.0.clone(),
         }
+    }
+}
+
+/// Middleware for adding some shareable value to [request extensions].
+///
+/// See [Sharing state with handlers](index.html#sharing-state-with-handlers)
+/// for more details.
+///
+/// [request extensions]: https://docs.rs/http/latest/http/struct.Extensions.html
+#[derive(Clone, Copy, Debug)]
+pub struct AddExtension<S, T> {
+    pub(crate) inner: S,
+    pub(crate) value: T,
+}
+
+impl<ResBody, S, T> Service<Request<ResBody>> for AddExtension<S, T>
+where
+    S: Service<Request<ResBody>>,
+    T: Clone + Send + Sync + 'static,
+{
+    type Response = S::Response;
+    type Error = S::Error;
+    type Future = S::Future;
+
+    #[inline]
+    fn poll_ready(&mut self, cx: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
+        self.inner.poll_ready(cx)
+    }
+
+    fn call(&mut self, mut req: Request<ResBody>) -> Self::Future {
+        req.extensions_mut().insert(self.value.clone());
+        self.inner.call(req)
     }
 }
