+++ axum/src/lib.rs
@@ -392,7 +392,6 @@
 #[macro_use]
 pub(crate) mod macros;
 
-mod add_extension;
 mod extension;
 #[cfg(feature = "json")]
 mod json;
@@ -409,7 +408,6 @@ pub mod routing;
 #[cfg(test)]
 mod test_helpers;
 
-pub use add_extension::AddExtension;
 #[doc(no_inline)]
 pub use async_trait::async_trait;
 #[cfg(feature = "headers")]
